+++ test/cases/cache-filesystem-js/webpack.config.js
@@ -11,7 +11,7 @@ module.exports = {
 
   output: {
     path: path.join(__dirname, 'dist/'),
-    clean: true,
+    clean: false,
   },
 
   plugins: [