+++ test/integration.test.js
@@ -31,8 +31,8 @@ describe('cache tests', () => {
   //   await new Promise((r) => setTimeout(r, 500));
   // });
 
-  // NOTE: use only single test with filesystem cache to avoid the error: 'PersistentCache is already registered'
-  //test('filesystem display stats', () => stdoutContain('cache-filesystem-display-stats', 'compiled successfully'));
+  test('filesystem, display stats', () => stdoutContain('cache-filesystem-display-stats', 'compiled successfully'));
+  test('filesystem, multiple config', () => compareFiles('cache-filesystem-multi-config'));
 
   test('filesystem-js-runs_n1', () => compareFilesRuns('cache-filesystem-js', false, 1));