+++ test/manual/cache-filesystem-multi-config/src/index.js
@@ -1,0 +1,3 @@
+import img from './iphone.png';
+
+console.log('>> index.js', img);